# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "TransDemo API",
  "description": "Backend réplica (demo) de **TransDemo S.R.L.**: venta de pasajes de bus, rutas e itinerarios, carga y encomiendas, puerta a puerta y centro de ayuda.\n\nFechas en hora de Bolivia (America/La_Paz) y montos en bolivianos (Bs).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/faqs": {
   "get": {
    "tags": [
     "Centro de ayuda"
    ],
    "summary": "Preguntas frecuentes por categoría",
    "operationId": "faqs_api_v1_faqs_get",
    "parameters": [
     {
      "name": "categoria",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "examples": [
        "pasajeros"
       ],
       "title": "Categoria"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FaqCategoriaOut"
         },
         "title": "Response Faqs Api V1 Faqs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/faqs/buscar": {
   "get": {
    "tags": [
     "Centro de ayuda"
    ],
    "summary": "Buscar en las preguntas frecuentes",
    "operationId": "buscar_api_v1_faqs_buscar_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 2,
       "examples": [
        "puedo llevar a mi perro"
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FaqBusquedaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/paginas/{slug}": {
   "get": {
    "tags": [
     "Centro de ayuda"
    ],
    "summary": "Una página de contenido",
    "operationId": "pagina_api_v1_paginas__slug__get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PaginaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "FaqBusquedaOut": {
    "properties": {
     "consulta": {
      "type": "string",
      "title": "Consulta"
     },
     "resultados": {
      "items": {
       "$ref": "#/components/schemas/FaqOut"
      },
      "type": "array",
      "title": "Resultados"
     },
     "mensaje": {
      "type": "string",
      "title": "Mensaje"
     }
    },
    "type": "object",
    "required": [
     "consulta",
     "resultados",
     "mensaje"
    ],
    "title": "FaqBusquedaOut"
   },
   "FaqCategoriaOut": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "codigo": {
      "type": "string",
      "title": "Codigo"
     },
     "nombre": {
      "type": "string",
      "title": "Nombre"
     },
     "preguntas": {
      "items": {
       "$ref": "#/components/schemas/FaqOut"
      },
      "type": "array",
      "title": "Preguntas"
     }
    },
    "type": "object",
    "required": [
     "codigo",
     "nombre",
     "preguntas"
    ],
    "title": "FaqCategoriaOut"
   },
...
   "FaqOut": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "pregunta": {
      "type": "string",
      "title": "Pregunta"
     },
     "respuesta": {
      "type": "string",
      "title": "Respuesta"
     },
     "respuesta_corta_voz": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Respuesta Corta Voz"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "pregunta",
     "respuesta",
     "respuesta_corta_voz"
    ],
    "title": "FaqOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PaginaOut": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "titulo": {
      "type": "string",
      "title": "Titulo"
     },
     "meta_descripcion": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Meta Descripcion"
     },
     "contenido_md": {
      "type": "string",
      "title": "Contenido Md"
     },
     "url_original": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url Original"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "titulo",
     "meta_descripcion",
     "contenido_md",
     "url_original"
    ],
    "title": "PaginaOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    }
   }
  }
 }
}